Reorganization + Solanum
- Solanum is now Gman
- Made separate namespaces for utils and controllers
