+++ HalfLifeOverhaul/HalfLifeAssets/AssetsToUse/hlmodels.manifest
@@ -1,9 +1,9 @@
 ManifestFileVersion: 0
-CRC: 1114055062
+CRC: 543222491
 Hashes:
   AssetFileHash:
     serializedVersion: 2
-    Hash: 8d1d26d6dd82276a21891a7de2a23b87
+    Hash: 6d7eafe6ef2062b577d2988242239464
   TypeTreeHash:
     serializedVersion: 2
     Hash: db59cb376d8c5d2b289f57e3f739454e
