+++ HalfLifeOverhaul/HalfLifeOverhaul.csproj
@@ -285,14 +285,15 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="IckyController.cs" />
-    <Compile Include="Logger.cs" />
+    <Compile Include="Controllers\GmanController.cs" />
+    <Compile Include="Controllers\IckyController.cs" />
+    <Compile Include="Util\Logger.cs" />
     <Compile Include="MainBehaviour.cs" />
     <Compile Include="MeshPatcher.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SkeletonSwapper.cs" />
-    <Compile Include="Utility.cs" />
-    <Compile Include="VortiguantController.cs" />
+    <Compile Include="Util\Utility.cs" />
+    <Compile Include="Controllers\VortiguantController.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="default-config.json" />
